pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ralphite"
version = "1.0.0"
description = "Ralphite local-first multi-agent CLI orchestrator"
requires-python = ">=3.13"
authors = [
    {name = "Alberto Mussali", email = "albertomussalih@gmail.com"},
]
license = {text = "PolyForm Internal Use 1.0.0 OR Commercial License"}
readme = "README.md"
keywords = ["agent", "orchestrator", "cli", "headless"]
classifiers = [
    "Programming Language :: Python :: 3.13",
    "License :: Other/Proprietary License",
]
dependencies = [
    "typer>=0.12,<1",
    "rich>=13.9,<14",
    "pydantic>=2.8,<3",
    "pyyaml>=6.0,<7",
]

[project.urls]
Homepage = "https://github.com/AlbertoMussali/Ralphite"
Documentation = "https://github.com/AlbertoMussali/Ralphite/blob/main/docs/index.md"
Repository = "https://github.com/AlbertoMussali/Ralphite"

[project.scripts]
ralphite = "ralphite.cli.main:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"ralphite.schemas" = ["json/*.json"]

[tool.ruff]
extend-exclude = ["build"]
